excepcion en productos

+++ src/main/java/com/example/utnfinaljava/services/ProductServiceImpl.java
@@ -7,8 +7,11 @@ import com.example.utnfinaljava.config.mappers.ProductMapper;
 import com.example.utnfinaljava.dtos.ProductDto;
 import com.example.utnfinaljava.entities.Product;
 import com.example.utnfinaljava.interfaces.ProductService;
+import com.example.utnfinaljava.repositories.OrderDetailRepository;
+import com.example.utnfinaljava.repositories.OrderRepository;
 import com.example.utnfinaljava.repositories.ProductRepository;
 import com.example.utnfinaljava.util.exceptions.NotExistException;
+import com.example.utnfinaljava.util.exceptions.PersistenceException;
 
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -20,6 +23,8 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
 
     private final ProductMapper productMapper;
+
+    private final OrderDetailRepository orderDetailRepository;
 
     @Override
     public List<ProductDto> getProducts() {
@@ -55,6 +60,9 @@ public class ProductServiceImpl implements ProductService {
         boolean notExist = !productRepository.existsById(id);
         if(notExist){
             throw new NotExistException("El producto ingresado no existe");
+        }
+        if(orderDetailRepository.existsByProductId(id)){
+            throw new PersistenceException("No se puede borrar un producto que este integrado en un pedido");
         }
         this.productRepository.deleteById(id);
     }
